+++ spalloc/_version.py
